+++ src/checks_intf.ml
@@ -3,24 +3,26 @@ open! Import
 type empty = |
 
 module type S = sig
+  type io
+
   module Stat : sig
-    val run : root:string -> unit
+    val run : io:io -> root:string -> unit
     (** Read basic metrics from an existing store. *)
 
-    val term : (unit -> unit) Cmdliner.Term.t
+    val term : io:io -> (unit -> unit) Cmdliner.Term.t
     (** A pre-packaged [Cmdliner] term for executing {!run}. *)
   end
 
   module Integrity_check : sig
-    val run : root:string -> unit
+    val run : io:io -> root:string -> unit
     (** Check that the integrity invariants of a store are preserved, and
         display any broken invariants. *)
 
-    val term : (unit -> unit) Cmdliner.Term.t
+    val term : io:io -> (unit -> unit) Cmdliner.Term.t
     (** A pre-packaged [Cmdliner] term for executing {!run}. *)
   end
 
-  val cli : unit -> empty
+  val cli : io:io -> unit -> empty
   (** Run a [Cmdliner] binary containing tools for running offline integrity
       checks. *)
 end
@@ -38,5 +40,6 @@ module type Checks = sig
   module type S = S
   module type Platform_args = Platform_args
 
-  module Make (K : Data.Key) (V : Data.Value) (_ : Platform_args) : S
+  module Make (K : Data.Key) (V : Data.Value) (P : Platform_args) :
+    S with type io = P.IO.io
 end