+++ test/cli/index_fsck.ml
@@ -7,4 +7,4 @@ module Index =
     (Index.Key.String_fixed (Size)) (Index.Value.String_fixed (Size))
     (Index.Cache.Noop)
 
-let () = match Index.Checks.cli () with _ -> .
+let () = match Index.Checks.cli ~io:() () with _ -> .