+++ CHANGES.md
@@ -6,6 +6,10 @@
 - Update to cmdliner.1.1.0 (#382, @MisterDA)
 - Mirage support: optional dependency to unix (#396, @art-w)
 
+## Added
+
+- Add `index.eio` with Eio backend (#397, @art-w)
+
 # 1.6.2 (2023-06-06)
 
 ## Changed